backend/requirements.txt
alembic==1.9.3
anyio==3.6.2
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.1
ecdsa==0.18.0
fastapi==0.91.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.2
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.4
PyJWT==2.6.0
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.3
starlette==0.24.0
typing_extensions==4.4.0
uvicorn==0.20.0
